Adding the logic of getting the height of header and then assigning to main margin top
