+++ scripts/utils.js
@@ -2,18 +2,3 @@ export default function getPathSegments() {
   return window.location.pathname.split('/')
     .filter((segment) => segment);
 }
-
-export function enableAnimationOnScroll() {
-  const observer = new IntersectionObserver((entries) => {
-    entries.forEach((entry) => {
-      if (entry.isIntersecting) {
-        entry.target.classList.add('visible');
-      } else {
-        entry.target.classList.remove('visible'); // remove class when out of view
-      }
-    });
-  }, { threshold: 0.1 }); // Trigger when 10% visible
-
-  const elements = document.querySelectorAll('.animate-on-scroll');
-  elements.forEach((el) => observer.observe(el));
-}
